세션 점유 종료 시 websocket event type 변경

+++ src/main/java/com/seatly/seatly/service/SessionService.java
@@ -77,23 +77,31 @@ public class SessionService {
     if (!isAdmin && !id.equals(redisService.getSessionIdByUserId(userId))) {
       throw new ForbiddenException("사용자 정보와 세션 아이디가 일치하지 않습니다.");
     }
+
     storeService.findById(id).ifPresent(session -> {
+      WebSocketEventType eventType;
       Long studyCafeId = session.getSeat().getStudyCafe().getId();
       Long seatId = session.getSeat().getId();
       Long sessionUserId = session.getUser().getId();
 
-      // TimePass 차감: startTime이 존재하면 이용 시간만큼 leftTime 차감
-      if (session.getStartTime() != null) {
-        long usedSeconds = Duration.between(session.getStartTime(), Util.now()).getSeconds();
-        UserTimePassId timePassId = new UserTimePassId(studyCafeId, sessionUserId);
-        UserTimePass timePass = userTimePassStoreService.findById(timePassId);
-        if (timePass != null) {
-          long newLeftTime = timePass.getLeftTime() - usedSeconds;
-          if (newLeftTime <= 0) {
-            userTimePassStoreService.deleteByUserIdAndStudyCafeId(sessionUserId, studyCafeId);
-          } else {
-            timePass.setLeftTime(newLeftTime);
-            userTimePassStoreService.save(timePass);
+      if (session.getStatus().equals(SessionStatus.ASSIGNED)) {
+        eventType = WebSocketEventType.SEAT_HOLD_RELEASED;
+      } else {
+        eventType = WebSocketEventType.SEAT_USAGE_FINISHED;
+
+        // TimePass 차감: startTime이 존재하면 이용 시간만큼 leftTime 차감
+        if (session.getStartTime() != null) {
+          long usedSeconds = Duration.between(session.getStartTime(), Util.now()).getSeconds();
+          UserTimePassId timePassId = new UserTimePassId(studyCafeId, sessionUserId);
+          UserTimePass timePass = userTimePassStoreService.findById(timePassId);
+          if (timePass != null) {
+            long newLeftTime = timePass.getLeftTime() - usedSeconds;
+            if (newLeftTime <= 0) {
+              userTimePassStoreService.deleteByUserIdAndStudyCafeId(sessionUserId, studyCafeId);
+            } else {
+              timePass.setLeftTime(newLeftTime);
+              userTimePassStoreService.save(timePass);
+            }
           }
         }
       }
@@ -102,9 +110,9 @@ public class SessionService {
       storeService.delete(session);
 
       // Global - 좌석 이용 종료 알림 전송
-      sendWebSocketEventGlobal(studyCafeId, seatId, WebSocketEventType.SEAT_USAGE_FINISHED);
+      sendWebSocketEventGlobal(studyCafeId, seatId, eventType);
       // User - 좌석 이용 종료 알림 전송
-      sendWebSocketEventUser(sessionUserId, seatId, WebSocketEventType.SEAT_USAGE_FINISHED);
+      sendWebSocketEventUser(sessionUserId, seatId, eventType);
     });
   }
